Smaller clippy fixes
- remove redundant calls to .into()
- remove redundant references
- remove calls to .clone() on types that implement Copy trait
- remove redundant lifetime annotations
- switch to is_empty instead of len() check
- switch to matches! macro
- collapse if-else blocks
- replace Option.map(...).flatten() with and_then(...)

+++ rust/ffi/src/ledger/datastore/lmdb/block_store.rs
@@ -134,7 +134,7 @@ pub unsafe extern "C" fn rsn_lmdb_block_store_count(
     handle: *mut LmdbBlockStoreHandle,
     txn: *mut TransactionHandle,
 ) -> u64 {
-    (*handle).0.count((*txn).as_txn()) as u64
+    (*handle).0.count((*txn).as_txn())
 }
 
 #[no_mangle]
